+++ packages/tasks-with-pinia/src/mount.ts
@@ -3,18 +3,18 @@ import { createPinia } from "pinia";
 import { type MountedFn } from "@efanworks/tasks-types";
 import Tasks from "./components/Tasks/index.vue";
 
-export type RoutePath = "tasks";
-
 const routes = {
   tasks: Tasks,
 };
 
-export const mount: MountedFn<RoutePath> = (root, path = "tasks") => {
+type RoutePath = keyof typeof routes;
+
+export const mount: MountedFn = (root, path = "tasks") => {
   if (!root) {
     throw new Error("Can not find mounted element");
   }
   const pinia = createPinia();
-  const app = createApp(routes[path]);
+  const app = createApp(routes[path as RoutePath]);
   app.use(pinia);
   app.mount(root);
 };