+++ packages/tasks-with-pinia/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @efanworks/tasks-with-pinia
 
+## 1.0.3
+
+### Patch Changes
+
+- e30d1bb: 新增tasks-with-mobx package
+- Updated dependencies [e30d1bb]
+  - @efanworks/tasks-api@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes