+++ packages/tasks-with-zustand/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @efanworks/tasks-with-zustand
 
+## 1.0.5
+
+### Patch Changes
+
+- 056e5c4: 移除不存在的组件导入 Counter
+
+## 1.0.4
+
+### Patch Changes
+
+- e30d1bb: 新增tasks-with-mobx package
+- Updated dependencies [e30d1bb]
+  - @efanworks/tasks-api@1.0.3
+
 ## 1.0.3
 
 ### Patch Changes